🦄 refactor: logic
